Moved arrangements section to its own component and fixed bug with elements 'flashing'/changing position
